refactor: update import path for ConvertSchema and rename to convertSchema
