+++ src/apps/gateway/videos/index.ts
@@ -1,17 +1,16 @@
 import express, { Router } from 'express';
 import { validateRequest } from 'src/utils/validator';
-import { ConvertRequest, ConvertSchema } from '../../../services/videos/convert/schema';
 import { crawlHandler } from './routes/crawl';
 import { CrawlRequest, CrawlSchema } from './routes/crawl/schema';
 import { fixVideosDuration } from './routes/fix-videos-duration';
 import { fixVideosThumbnail } from './routes/fix-videos-thumbnail';
 import { streamToStorage } from './routes/stream-to-storage';
 import { HasuraWebhookRequest, hasuraWebhookSchema } from 'src/schema/hasura';
+import { ConvertRequest, convertSchema } from 'src/schema/videos/convert';
 
 const videosRouter: Router = express.Router();
 
-// TODO: refactor the ConvertSchema from the old legacy code
-videosRouter.post('/convert', validateRequest<ConvertRequest>(ConvertSchema), streamToStorage);
+videosRouter.post('/convert', validateRequest<ConvertRequest>(convertSchema), streamToStorage);
 videosRouter.post(
   '/fix-videos-duration',
   validateRequest<HasuraWebhookRequest>(hasuraWebhookSchema),
